replace vec with boundedvec issue #54

+++ runtime/mrc/src/lib.rs
@@ -457,7 +457,7 @@ impl pallet_collator_selection::Config for Runtime {
 }
 
 parameter_types! {
-	pub const MaxMessagesPerBlock: u16 = 7000;
+	pub const MaxMessagesPerBlock: u32 = 7000;
 	pub const MaxMessageSizeInBytes: u32 = 1024 * 50; // 50K
 }
